All sections on the same page
Today I joined all the different pages on the same html file. I combined the css and js to have 1 whole project in 1 place.

I worked on the transitions (made the first work), added commentaries for the code, and added some code that will be useful for my text entry page. 

I also made a second js file to put the code for the transition so it would be more clear.