last bit chunks of the code added
The website is now fully functional. I added the code missing to make the text entry page.
I also added the array containing the links that are useful at the end, and the bits of code to fill it as the user go through the website.
Finally, I added the code to spam the user with popups at the end.