Final Update
I decided to add a timer of 8 seconds before the spamming at the end start, so the person have the time to read what is said on the last page before getting destroyed.

Project is done! I like the website pretty much. I kind of had a basis in javascript, but I find jQuery is very useful, and easier to use in lots of situations. Also, the libraries are very, very useful as they can do very complex stuff very easily. This class was the most useful web class for me, and I'm glad I took it!